feature-database-setup/requirements.txt
sqlmodel>=0.0.8
pydantic>=1.10.0
python-dotenv>=0.21.0
SQLAlchemy>=1.4.0
psycopg2-binary>=2.9.3  # For PostgreSQL support
python-jose[cryptography]>=3.3.0  # For JWT authentication
passlib[bcrypt]>=1.7.4  # For password hashing
python-multipart>=0.0.5  # For form data handling
email-validator>=1.3.0  # For email validation
python-dateutil>=2.8.2  # For date handling
pytest>=7.0.0  # For testing
pytest-cov>=3.0.0  # For test coverage
black>=22.8.0  # For code formatting
isort>=5.10.1  # For import sorting
mypy>=0.982  # For static type checking
httpx>=0.23.0  # For HTTP client
pandas>=1.5.0  # For data manipulation
numpy>=1.23.0  # For numerical operations
scikit-learn>=1.1.0  # For machine learning
fastapi>=0.85.0  # For API development
uvicorn>=0.19.0  # ASGI server for FastAPI
alembic>=1.8.1  # For database migrations
pydantic[email]>=1.10.0  # For email validation
